move mail file
