+++ app/modules/booking/email-helpers.ts
@@ -1,10 +1,10 @@
 import { bookingUpdatesTemplateString } from "~/emails/bookings-updates-template";
+import { sendEmail } from "~/emails/mail.server";
 import type { BookingForEmail } from "~/emails/types";
 import { getDateTimeFormatFromHints } from "~/utils/client-hints";
 import { getTimeRemainingMessage } from "~/utils/date-fns";
 import { SERVER_URL } from "~/utils/env";
 import { ShelfError } from "~/utils/error";
-import { sendEmail } from "~/utils/mail.server";
 import type { ClientHint } from "./types";
 
 /**
