+++ app/routes/api+/stripe-webhook.ts
@@ -3,10 +3,10 @@ import type { ActionFunctionArgs } from "@remix-run/node";
 import { json } from "@remix-run/node";
 import type Stripe from "stripe";
 import { db } from "~/database/db.server";
+import { sendEmail } from "~/emails/mail.server";
 import { trialEndsSoonText } from "~/emails/stripe/trial-ends-soon";
 import { ShelfError, makeShelfError } from "~/utils/error";
 import { error } from "~/utils/http.server";
-import { sendEmail } from "~/utils/mail.server";
 import {
   fetchStripeSubscription,
   getDataFromStripeEvent,
